+++ src/angular.js
@@ -2,7 +2,7 @@
 /*global inlineAttachment: false, angular: false */
 
 import Utils from "./utils";
-import InputInlineAttachment from "./input.inline-attachment";
+import InputInlineAttachment from "./input";
 
 let module = ng.module('inlineattachment', []);
 let directiveName = 'inlineattachment';
